+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 v1.3.1 (2021)
 - Added new module RandRoute
-- 
+- Bug fix for Orbitones trails
 
 v1.2.4 (3-05-2021)
 - Bug fix in Orbitones